Fresh-fold-Backend/requirements.txt
Django>=5.0,<6
djangorestframework>=3.15
django-cors-headers>=4.3
python-dotenv>=1.0
PyMySQL>=1.1
reportlab>=4.5
